backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
pydantic-ai>=0.0.14
python-dotenv>=1.0.0

# AI & LLM
google-generativeai>=0.3.2
openai>=1.12.0

# Database
pymongo>=4.6.1
motor>=3.3.2

# Google Sheets
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
gspread>=6.0.2

# Observability
logfire>=0.36.0

# Utilities
httpx>=0.27.2
python-multipart>=0.0.6
